:zap: (366): faster test by auto-using mocks
